% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactBeta.R
\name{exactBeta}
\alias{exactBeta}
\title{Calculate lower and upper CI of a given quantile using exact method, based on beta distribution}
\usage{
exactBeta(n, q, alpha)
}
\arguments{
\item{n}{sample size}

\item{q}{quantile}

\item{alpha}{desired significance level}
}
\value{
a list of the lower and upper confidence limit of the quantiles.Values are between [0,1]
\item{u1}{lower confidence limit of the quantile}
\item{u2}{upper confidence limit of the quantile}
}
\description{
Calculate lower and upper CI of a given quantile using exact method, based on beta distribution
}
\examples{
QuantileNPCI:::exactBeta(25, 0.5, 0.05)

}
