% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood.R
\docType{data}
\name{flood}
\alias{flood}
\title{The flood rate of Feature River and Blackstone River.}
\format{A data frame with 96 rows and 3 variables:
\describe{
  \item{loc}{River name}
  \item{year}{year of the record}
  \item{discharge}{flood discharge rate}
}}
\usage{
flood
}
\description{
A dataset containing the flood rate data, as presented in Hutson 1999 paper.
Original source: Pericchi and Rodreiguez-Iturbe (1995)
}
\keyword{datasets}
